Move commands out of internal package and fix my bad formatting with x5424
